house-price-api/requirements.txt
anyio==3.5.0
asyncpg==0.25.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
coverage==6.3.2
databases==0.5.5
fastapi==0.74.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
joblib==1.1.0
loguru==0.5.3
numpy==1.22.2
packaging==21.3
pandas==1.4.1
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
requests==2.27.1
scikit-learn==0.24.1
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
threadpoolctl==3.1.0
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
uvicorn==0.13.3
